requirements.txt
beautifulsoup4
pandas
Pillow
pyautogui
pynput
pyperclip
requests
scrapetube

famgz-utils @ git+https://github.com/famgz/famgz-utils.git@main